feat(sdk): implement Effect-TS based SDK and utility functions
