+++ packages/studiocms/src/lib/effects/index.ts
@@ -1,0 +1,1 @@
+export { convertToVanilla } from './convertToVanilla.js';
